fix parse errors for non_repeaters & max_repetitions: aligment errors
git-svn-id: http://svn.php.net/repository/php/php-src/php/php-src/trunk@308682 c90b9560-bf6c-de11-be94-00142212c4b1

+++ ext/snmp/snmp.c
@@ -420,8 +420,8 @@ struct objid_query {
 	int count;
 	int offset;
 	int step;
-	int non_repeaters;
-	int max_repetitions;
+	long non_repeaters;
+	long max_repetitions;
 	int valueretrieval;
 	int array_output;
 	snmpobjarg *vars;
